Added configuration file for listing servers. TV now tries to connect to all servers on the list  (preferably simultaneously). Other clients now connect to the first available server on the list.
Server address given on command line is added to the start of the list, if both given.

If no address on command line, server address given in environment variables is added to the start of the list, if both given.

Removed default servers and ports from java code (except for localhost:61910).

Possibly other staff as well.